Move state component under igniteui-angular-elements, so that it does not appear in the angular api.
