+++ projects/igniteui-angular-elements/src/lib/state.component.ts
@@ -1,0 +1,34 @@
+import { Input, Component, EnvironmentInjector, Inject, Injector, ViewContainerRef } from '@angular/core';
+import { GridServiceType, IGX_GRID_SERVICE_BASE, IGridStateOptions, IgxGridStateDirective } from 'igniteui-angular';
+
+/* blazorElement */
+/* mustUseNGParentAnchor */
+/* wcElementTag: igc-grid-state */
+/* blazorIndirectRender */
+/* contentParent: GridBaseDirective */
+/* jsonAPIManageItemInMarkup */
+
+/**
+ * State component description
+ * @igxParent IgxGridComponent, IgxTreeGridComponent, IgxHierarchicalGridComponent, IgxPivotGridComponent, *
+ */
+@Component({
+    selector: 'igx-grid-state',
+    template: ``,
+    standalone: true
+})
+export class IgxGridStateComponent extends IgxGridStateDirective {
+
+    constructor(
+        @Inject(IGX_GRID_SERVICE_BASE) private api: GridServiceType,
+        protected override viewRef: ViewContainerRef, protected  override envInjector: EnvironmentInjector,
+        protected override injector: Injector,
+        ) {
+            super(api.grid, viewRef, envInjector, injector);
+        }
+
+        @Input()
+        public override get options(): IGridStateOptions {
+           return this._options;
+        }
+}
