+++ src/app/grid-state/grid-state.component.ts
@@ -51,16 +51,13 @@ interface GridState {
     styleUrls: ['./grid-state.component.scss'],
     templateUrl: './grid-state.component.html',
     standalone: true,
-    imports: [IgxExpansionPanelComponent, IgxExpansionPanelHeaderComponent, IgxGridStateComponent, IgxExpansionPanelTitleDirective, NgIf, IgxExpansionPanelIconDirective, IgxExpansionPanelBodyComponent, NgTemplateOutlet, IgxGridComponent, IgxGridStateDirective, IgxGridToolbarComponent, IgxGridToolbarActionsComponent, IgxGridToolbarPinningComponent, IgxGridToolbarHidingComponent, IgxGridToolbarAdvancedFilteringComponent, NgFor, IgxColumnComponent, IgxGridDetailTemplateDirective, IgxPaginatorComponent, IgxTooltipDirective, IgxColumnGroupComponent, IgxHierarchicalGridComponent, IgxRowIslandComponent, IgxTreeGridComponent, RouterLink, IgxTooltipTargetDirective, IgxIconComponent, IgxSwitchComponent, FormsModule, IgxButtonDirective, IgxCellTemplateDirective]
+    imports: [IgxExpansionPanelComponent, IgxExpansionPanelHeaderComponent, IgxExpansionPanelTitleDirective, NgIf, IgxExpansionPanelIconDirective, IgxExpansionPanelBodyComponent, NgTemplateOutlet, IgxGridComponent, IgxGridStateDirective, IgxGridToolbarComponent, IgxGridToolbarActionsComponent, IgxGridToolbarPinningComponent, IgxGridToolbarHidingComponent, IgxGridToolbarAdvancedFilteringComponent, NgFor, IgxColumnComponent, IgxGridDetailTemplateDirective, IgxPaginatorComponent, IgxTooltipDirective, IgxColumnGroupComponent, IgxHierarchicalGridComponent, IgxRowIslandComponent, IgxTreeGridComponent, RouterLink, IgxTooltipTargetDirective, IgxIconComponent, IgxSwitchComponent, FormsModule, IgxButtonDirective, IgxCellTemplateDirective]
 })
 export class GridSaveStateComponent implements OnInit {
     @ViewChild(IgxExpansionPanelComponent, { static: true })
     private igxExpansionPanel: IgxExpansionPanelComponent;
     @ViewChildren(IgxGridStateDirective)
     private state: QueryList<IgxGridStateDirective>;
-
-    @ViewChild(IgxGridStateComponent)
-    private stateComponent: IgxGridStateComponent;
 
     public localData = employeesData;
     public hierData = this.generateDataUneven(10, 3);
@@ -137,7 +134,7 @@ export class GridSaveStateComponent implements OnInit {
 
     public getContext(grid: GridType) {
         if (this.state) {
-        const stateDirective = this.state.find(st => st.grid.id === grid.id) || this.stateComponent;
+        const stateDirective = this.state.find(st => st.grid.id === grid.id);
         return { $implicit: grid, stateDirective};
         }
         return { $implicit: grid };
